retrieve all orders for a specific user done

+++ src/app/modules/user/user.validation.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 const orderValidationSchema = z.object({
     productName: z.string().min(1, { message: 'Product name is required' }),
     price: z.number().min(0, { message: 'Price must be a non-negative number' }),
-    quantity: z.number().min(1, { message: 'Quantity must be at least 1' }),
+    quantity: z.number().int().min(1, { message: 'Quantity must be at least 1' }),
 });
 
 const userValidationSchema = z.object({
